+++ SmallServerAdmin/global.php
@@ -3,7 +3,7 @@
 require_once 'config.php';
 # WebForms.PHP Application
 require_once $_SERVER['DOCUMENT_ROOT'].'/Libs/Nemiro/App.php';
-# SmallServerAdmin config
+# SmallServerAdmin default config
 require_once 'ssa.config.php';
 
 # import and init application class
@@ -19,6 +19,28 @@ App::AddHandler('Application_PageCreated');
 # get from https://github.com/alekseynemiro/Nemiro.Data.PHP
 # App::IncludeFile('~/Nemiro/Data');
 
+# config for selected server
+if (isset($_COOKIE['currentServer']) && $_COOKIE['currentServer'] != '')
+{
+  if (file_exists(\Nemiro\Server::MapPath('~/servers/'.$_COOKIE['currentServer'].'.php')))
+  {
+    # config found, include
+    \Nemiro\Console::Info('Server config: '.$_COOKIE['currentServer'].'.php.');
+    require_once \Nemiro\Server::MapPath('~/servers/'.$_COOKIE['currentServer'].'.php');
+    # fix client-side config
+    $config['client']['ServerAddress'] = $config['ssh_host'];
+    $config['client']['ServerName'] = $config['server_name'];
+  }
+  else
+  {
+    # config not found, remove from cookies
+    \Nemiro\Console::Warning('Server config "'.$_COOKIE['currentServer'].'.php" not found.');
+    unset($_COOKIE['currentServer']);
+    setcookie('currentServer', null, -1, '/');
+  }
+}
+
+# language
 $CurrentLang = (isset($_COOKIE['lang']) ? $_COOKIE['lang'] : PAGE_DEFAULT_CULTURE);
 
 # application event handlers