+++ SmallServerAdmin/logout.php
@@ -1,0 +1,3 @@
+<?php
+  header('Location: //log:out@'.$_SERVER['HTTP_HOST']);
+?>