+++ SmallServerAdmin/servers/192.168.56.140.php
@@ -1,0 +1,18 @@
+<?php
+// server config
+
+// !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+// WARNING:!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+// DO NOT DECLARE $config!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+// !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+
+$config['server_name'] = 'Local Server #140';
+$config['server_description'] = 'Test server / Debian 8.2 x64 / 1 CPU / 256 RAM / HDD 20 Gb';
+$config['server_disabled'] = TRUE;
+
+// ssh
+$config['ssh_host'] = '192.168.56.140';
+$config['ssh_port'] = '22';
+$config['ssh_user'] = 'ssa';
+$config['ssh_password'] = '6fYCiai8myTX';
+$config['ssh_required_password'] = TRUE;