+++ SmallServerAdmin/settings/api.php
@@ -29,16 +29,8 @@ namespace Api
   class Settings
   {
 
-    /**
-     * SSH client.
-     * 
-     * @var SSH
-     */
-    private $SshClient = NULL;
-
     function __construct()
     {
-      $this->SshClient = new SSH();
     }
 
     #region Update
@@ -100,45 +92,34 @@ namespace Api
       else
       {
         // current version is actual
-        return ['NeedUpdate' => FALSE];
+        return [
+          'NeedUpdate' => FALSE, 
+          'LatestVersion' => $latestVersion, 
+          'CurrentVersion' => $currentVersion
+        ];
       }
     }
 
     public function Update($data)
     {
-      $currentVersion = file_get_contents(\Nemiro\Server::MapPath('~/.version'));
-
-      if ($currentVersion === FALSE)
+      if (is_file(\Nemiro\Server::MapPath('~/settings/update.sh')) === FALSE)
       {
-        throw new \ErrorException('Cannot get current version from "'.\Nemiro\Server::MapPath('~/.version').'"');
+        throw new \ErrorException('File "'.\Nemiro\Server::MapPath('~/settings/update.sh').' not found.');
       }
 
-      if (preg_match('/\d+\.\d+\.\d+/', $currentVersion, $matches))
-      {
-        $currentVersion = $matches[0];
-      }
-      else
-      {
-        throw new \ErrorException('Cannot parse version number from string "'.$currentVersion.'"');
-      }
-
+      $sshClient = new SSH();
       $path = \Nemiro\Server::MapPath('~/');
-      $top_path = dirname($path);
-      // export files
-      $command = 'svn export https://github.com/adminstock/ssa.git/trunk/SmallServerAdmin /tmp/ssa && ';
-      // full backup
-      $command .= 'tar -zcf /var/backups/ssa-webpanel-v'.$currentVersion.'.tar.gz '.$path.' && ';
-      // backup ssa.config.php
-      $command .= 'mv '.\Nemiro\Server::MapPath('~/ssa.config.php').' '.$top_path.'/ssa.config.backup.php && ';
-      // remove old version
-      $command .= 'rm -r '.$path.' && ';
-      // set new version
-      $command .= 'mv /tmp/ssa '.$path.' && ';
-      // restore ssa.config.php
-      $command .= 'rm '.\Nemiro\Server::MapPath('~/ssa.config.php').' && ';
-      $command .= 'mv '.$top_path.'/ssa.config.backup.php '.\Nemiro\Server::MapPath('~/ssa.config.php');
+      $scriptPath = \Nemiro\Server::MapPath('~/settings/update.sh');
 
-      $shell_result = $this->SshClient->Execute($command);
+      $command = 'sudo bash -c \'';
+      $command .= 'updatePath="$(mktemp --dry-run /tmp/XXXXX.update.sh)"; ';
+      $command .= 'cp '.$scriptPath.' \$updatePath && ';
+      $command .= 'chmod +x \$updatePath && ';
+      $command .= '\$updatePath "'.$path.'" && ';
+      $command .= 'rm \$updatePath';
+      $command .= '\'';
+
+      $shell_result = $sshClient->Execute($command);
 
       if ($shell_result->Error != '')
       {
@@ -146,6 +127,50 @@ namespace Api
       }
 
       return ['Success' => TRUE];
+    }
+
+    #endregion
+    #region Servers
+
+    /**
+     * Returs servers list.
+     */
+    public function GetServers()
+    {
+      global $config;
+      $servers = [];
+
+      // default server
+      $servers[] = $this->GetServerInfo(\Nemiro\Server::MapPath('~/ssa.config.php'), TRUE);
+
+      // get others servers
+      if (is_dir(\Nemiro\Server::MapPath('~/servers')))
+      {
+        foreach (scandir(\Nemiro\Server::MapPath('~/servers')) as $file) 
+        {
+          // NOTE: ssa.config.php - reserved for default (root) config
+          if ($file == '.' || $file == '..' || $file == 'ssa.config.php' || pathinfo($file, PATHINFO_EXTENSION) != 'php') { continue; }
+          $servers[] = $this->GetServerInfo(\Nemiro\Server::MapPath('~/servers/'.$file), FALSE);
+        }
+
+        usort($servers, function($a, $b) { return strcmp($a->Name, $b->Name); });
+      }
+
+      return $servers;
+    }
+
+    private function GetServerInfo($path, $default)
+    {
+      require $path;
+
+      return [
+        'Address' => $config['ssh_host'], 
+        'Name' => $config['server_name'], 
+        'Description' => $config['server_description'], 
+        'Config' => basename($path, '.php'),
+        'IsDefault' => $default,
+        'Disabled' => ($default ? FALSE : (bool)$config['server_disabled'])
+      ];
     }
 
     #endregion