+++ SmallServerAdmin/settings/servers.php
@@ -1,0 +1,14 @@
+<?php
+namespace Settings
+{
+
+  require_once '../global.php';
+
+  class Update extends \Nemiro\UI\Page
+  {
+
+  }
+
+  \Nemiro\App::Magic(__NAMESPACE__);
+
+}