+++ CHANGELOG.md
@@ -2,10 +2,20 @@
 
 All notable changes to SmallServerAdmin will be documented in this file.
 
-## [1.0.34-alpha] - unreleased
+## [1.0.37-alpha] - unreleased
 ### Added
 - @adminstock news widget;
 - New module - **settings** with update functions;
+- Logout button;
+- Support for multiple servers;
+
+### Changed
+- References to script and style file moved to `StaticIncludes` control;
+- The language selection interface changed;
+
+### Fixed
+- Fixed problem with setting the password to svn users, 
+  if the password file does not exist;
 
 ## [1.0.32-alpha] - 2016-04-27
 First alpha.