+++ SmallServerAdmin/settings/CHANGELOG.md
@@ -2,5 +2,12 @@
 
 All notable changes to this module will be documented in this file.
 
+## [1.0.1] - 2016-04-30
+### Added
+- Server Manager.
+
+### Changed
+- Restore servers folder after panel update.
+
 ## [1.0.0] - 2016-04-28
 The module was created by @alekseynemiro.