+++ SmallServerAdmin/svn/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this module will be documented in this file.
 
+## [1.0.2] - 2016-04-30
+### Fixed
+- Fixed problem with setting the password, if the password file does not exist:
+  https://github.com/adminstock/ssa/issues/11
+
 ## [1.0.1] - 2016-04-10
 ### Fixed
 - `$` escaping problem (*api.php*, method `SetAuthz`).